# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IncidentDesk",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/api/v1/investigations/natural": {
      "post": {
        "summary": "Interpret one sentence with Mistral; 200 clarification or 202 scoped investigation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 4000
                  },
                  "timezone": {
                    "type": "string"
                  }
                },
                "required": [
                  "text",
                  "timezone"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      }
    },
    "/api/v1/queries": {
      "post": {
        "summary": "Conversational assistant; answer, clarification or durable query job",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "history": {
                    "type": "array",
                    "maxItems": 12,
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000
                        }
                      },
                      "required": [
                        "role",
                        "content"
                      ]
                    }
                  },
                  "timezone": {
                    "type": "string"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "warehouse",
                      "investigation"
                    ]
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      },
      "get": {
        "summary": "Latest 50 owned business queries, filtered by current dataset access",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
    "/api/v1/queries/{id}": {
      "get": {
        "summary": "Persisted query progress, executed SQL, rows and errors",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/conversations": {
      "post": {
        "summary": "Idempotently create an owned persistent conversation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      },
      "get": {
        "summary": "Latest 50 owned conversations filtered by current source permissions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000
            }
          }
        ]
      }
    },
    "/api/v1/conversations/{id}": {
      "get": {
        "summary": "Owned conversation with latest 50 turns; before sequence cursor loads earlier turns",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/v1/conversations/{id}/turns": {
      "post": {
        "summary": "Append using server-owned history, expected revision and idempotency key; replay is safe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "warehouse",
                      "investigation"
                    ]
                  },
                  "timezone": {
                    "type": "string"
                  },
                  "revision": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "required": [
                  "text",
                  "revision"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      }
    },
    "/api/v1/datasets": {
      "get": {
        "summary": "Enabled datasets visible to current team memberships",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "summary": "Current identity and memberships",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
    "/api/v1/services": {
      "get": {
        "summary": "Accessible services",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
...
    "/api/v1/investigations/{id}": {
      "get": {
        "summary": "Investigation, evidence, action and freshness",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OIDC JWT"
      },
      "SourceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Source-Key"
      },
      "WorkerKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Worker-Key"
      }
    }
  }
}